+++ addons/hammerforge/face_data.gd
@@ -179,6 +179,7 @@ func to_dict() -> Dictionary:
 		"uv_offset": _encode_vec2(uv_offset),
 		"uv_rotation": uv_rotation,
 		"uv_format_version": 1,
+		"winding_version": 1,
 		"custom_uvs": _encode_vec2_array(custom_uvs),
 		"local_verts": _encode_vec3_array(local_verts),
 		"normal": _encode_vec3(normal),
@@ -400,7 +401,7 @@ func _compute_normal() -> void:
 	var a = local_verts[0]
 	var b = local_verts[1]
 	var c = local_verts[2]
-	var n = (b - a).cross(c - a)
+	var n = (c - a).cross(b - a)
 	if n.length() > 0.0001:
 		normal = n.normalized()
 	else: